requirements.txt
Flask~=2.1.3
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
werkzeug==2.1.2
flake8==5.0.4
PyJWT==2.4.0
Flask-JWT-Extended==4.4.4
coverage==6.4.4
SQLAlchemy==1.4.40
pytest-cov==3.0.0
pytest==7.1.2